moved MDXEditor files to global components folder, fixed styling for toolbar in editor so it flexes and jump down if dialog gets too small like in mobile. Some minor text changes.
